src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.2.2=h56e8100_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
filelock=3.13.1=py310haa95532_0
freetype=2.12.1=ha860e81_0
fsspec=2024.3.1=pyhca7485f_0
gmpy2=2.1.2=py310h7f96b67_0
huggingface_hub=0.22.2=pyhd8ed1ab_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py310haa95532_0
importlib-metadata=7.1.0=pyha770c72_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.3=py310haa95532_0
joblib=1.4.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py310h2bbff1b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
networkx=3.1=py310haa95532_0
nltk=3.8.1=pyhd8ed1ab_0
numpy=1.24.3=py310hdc03b94_0
numpy-base=1.24.3=py310h3caf3d7_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
packaging=23.2=py310haa95532_0
pillow=10.3.0=py310h2bbff1b_0
pip=23.3.1=py310haa95532_0
platformdirs=3.10.0=py310haa95532_0
pooch=1.7.0=py310haa95532_0
python=3.10.14=he1021f5_0
pytorch=2.3.0=py3.10_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py310h2bbff1b_0
regex=2023.10.3=py310h2bbff1b_0
requests=2.31.0=py310haa95532_1
safetensors=0.4.2=py310h0010962_0
scikit-learn=1.4.2=py310h4ed8f06_1
scipy=1.10.1=py310hb9afe5d_0
sentence-transformers=2.7.0=pyhd8ed1ab_0
sentencepiece=0.1.99=py310h59b6b97_0
setuptools=68.2.2=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
sympy=1.12=py310haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.15.1=py310h49fca51_0
torchaudio=2.3.0=pypi_0
torchvision=0.18.0=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
transformers=4.37.2=py310haa95532_0
typing-extensions=4.9.0=py310haa95532_1
typing_extensions=4.9.0=py310haa95532_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py310haa95532_0
xz=5.4.6=h8cc25b3_0
yaml=0.2.5=he774522_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
